+++ site/scenes/index.js
@@ -6,6 +6,7 @@
 //
 // Adding an animation is two edits: drop a scene module in this folder, then add it here.
 
+import * as longCut from './long-cut/index.js';
 import * as pitilessPit from './pitiless-pit/index.js';
 import * as moonOverTheDeep from './moon-over-the-deep/index.js';
 import * as roseFunnel from './rose-funnel/index.js';
@@ -13,6 +14,14 @@ import * as aboveTheFog from './above-the-fog/index.js';
 import * as grizzlyPeak from './grizzly-peak/index.js';
 import * as floatingBed from './floating-bed/index.js';
 
-export const scenes = [pitilessPit, moonOverTheDeep, roseFunnel, aboveTheFog, grizzlyPeak, floatingBed];
+export const scenes = [
+  longCut,
+  pitilessPit,
+  moonOverTheDeep,
+  roseFunnel,
+  aboveTheFog,
+  grizzlyPeak,
+  floatingBed,
+];
 
 export const findScene = (id) => scenes.find((scene) => scene.meta.id === id);